test: merge `auto_import_translation_functions` tests
