+++ specs/fixtures/basic_usage/pages/index.vue
@@ -21,11 +21,6 @@ const switchLocalePath = useSwitchLocalePath()
 const localeRoute = useLocaleRoute()
 const appConfig = useAppConfig()
 
-if (useNuxtApp().$config.public.i18n.experimental.autoImportTranslationFunctions) {
-  console.log(`[autoImportTranslationFunctions][default]: ${$t('welcome')}`)
-  console.log(`[autoImportTranslationFunctions][fr]: ${$t('welcome', 1, { locale: 'fr' })}`)
-}
-
 const category = ref({
   title: 'Kirby',
   slug: 'nintendo'
